# projects/tasks/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Task REST API
  version: 1.0.0
  description: >-
    A small framework-neutral task contract used by the applied REST project.
servers:
  - url: http://127.0.0.1:8000
paths:
  /health:
    get:
      operationId: getHealth
      summary: Check server readiness
      responses:
        "200":
          description: The server can answer requests.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
  /tasks:
    post:
      operationId: createTask
      summary: Create an incomplete task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTask"
      responses:
        "201":
          description: The created task.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/InvalidJson"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listTasks
      summary: List tasks
      parameters:
        - name: completed
          in: query
          required: false
          description: Filter by completion state.
          schema:
            type: boolean
      responses:
        "200":
          description: Tasks ordered by ID ascending.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /tasks/{taskId}:
    parameters:
      - $ref: "#/components/parameters/TaskId"
    get:
      operationId: getTask
      summary: Get one task
      responses:
        "200":
          description: The requested task.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      operationId: updateTask
      summary: Update a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTask"
      responses:
        "200":
          description: The updated task.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/InvalidJson"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteTask
      summary: Delete a task
      responses:
        "204":
          description: The task was deleted. The response has no body.
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  parameters:
    TaskId:
      name: taskId
      in: path
      required: true
      description: A positive task ID.
      schema:
        type: integer
        minimum: 1
  schemas:
    Health:
      type: object
      required:
        - status
      properties:
        status:
          const: ok
      additionalProperties: false
      examples:
        - status: ok
    Task:
      type: object
      required:
        - id
        - title
        - completed
      properties:
        id:
          type: integer
          minimum: 1
        title:
          type: string
          minLength: 1
          maxLength: 120
        completed:
          type: boolean
      additionalProperties: false
      examples:
        - id: 1
          title: Learn REST
          completed: false
    CreateTask:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
      additionalProperties: false
      examples:
        - title: Learn REST
    UpdateTask:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
        completed:
          type: boolean
      additionalProperties: false
      examples:
        - completed: true
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - invalid_json
                - payload_too_large
                - unsupported_media_type
                - not_found
                - method_not_allowed
                - validation_error
                - internal_error
            message:
              type: string
              minLength: 1
            details:
              type: object
              additionalProperties: true
          additionalProperties: false
      additionalProperties: false
  responses:
    InvalidJson:
      description: The request body could not be decoded as JSON.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            invalidJson:
              value:
                error:
                  code: invalid_json
                  message: request body must be valid JSON
    PayloadTooLarge:
      description: The request body exceeds the server's size limit.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            payloadTooLarge:
              value:
                error:
                  code: payload_too_large
                  message: request body exceeds the 1 MiB limit
    UnsupportedMediaType:
      description: >-
        The request Content-Type is missing, is not application/json, or
        declares a non-UTF-8 charset.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            unsupportedMediaType:
              value:
                error:
                  code: unsupported_media_type
                  message: request Content-Type must be application/json
    NotFound:
      description: The task or route was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            taskNotFound:
              value:
                error:
                  code: not_found
                  message: task 99 was not found
    MethodNotAllowed:
      description: The path does not support the request method.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            methodNotAllowed:
              value:
                error:
                  code: method_not_allowed
                  message: method is not allowed for this path
    ValidationError:
      description: The decoded request violates shape or domain rules.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            invalidTitle:
              value:
                error:
                  code: validation_error
                  message: title must contain between 1 and 120 characters
                  details:
                    field: title
    InternalError:
      description: The server could not complete the request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            internalError:
              value:
                error:
                  code: internal_error
                  message: the server could not complete the request